WIP Resize Fields
