+++ unit_tests/actuator/CMakeLists.txt
@@ -7,7 +7,7 @@ target_sources(${utest_ex_name} PRIVATE
 )
 if(ENABLE_OPENFAST)
    target_sources(${utest_ex_name} PRIVATE
-      ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorDiskFAST.C
+      ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorSweptPointLocator.C
       ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorFASTGaussian.C
       ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorParsingFAST.C
    	  ${CMAKE_CURRENT_SOURCE_DIR}/UnitTestActuatorFunctorsFAST.C
